caza: patch sensors.hal.tof to use proper input device
This library looks for the /dev/input/eventX event which has the name
STM VL53L1 proximity sensor, and then uses the X to find the input
device as /sys/class/input/inputX, but event character devices do not
exactly match the input device numbering.

Patch the lib to look for /sys/class/input/fakeiX, and add a change in
kernel to create a symlink from fakeiX where X matches the eventX number
to the matching input device.

Change-Id: I6764bd81011b488eaa0c195285b607b611d0a271